fix(rendering): compute baseline and thus y offset correctly
I've tried to document what the different offsets and values mean in an
ASCII diagram in the code which has now been added. Note that this
currently contradicts the documentation in stb_truetype.h which I
explain with the difference in direction of y-coordinates for bitmap and
non-bitmap rendering target in stb_truetype.h

We also currently don't consider stbtt_GetGlyphBitmapBox which seems to
be unnecessary as it seems to return the same values as the offsets
stbtt_GetGlyphBitmap yields. This will have to be tested with other
fonts to confirm however.

Also update demo output to the version with proper offsets.