feat(buchstabensuppe): font fallback
By default we now try to fallback to rendering 0xFFFD if a glyph is
missing from a grapheme in all fonts.

Also introduces BS_RENDER_NO_FALLBACK flag to disable this behavior.

Implementation reuses bs_render_grapheme_append (recursively) which is
probably overkill, but is concise and prevents code duplication.