chore: port to meson
This means that we can finally build a shared library correctly without
to much hassle and will be able to use pkgsStatic as well (provided we
fix the dependencies upstream).

redo is nice, but, while its dependency tracking is better, you end up
scripting a lot in a lot of different places.